requirements.txt
##################################################################
# core python modules
##################################################################
Django==1.11.1
channels==0.17.3
django_environ==0.4.3


##################################################################
# heroku hosting
##################################################################
dj-database-url==0.4.1
whitenoise==3.2
psycopg2
hiredis==0.2.0
django_redis
django_storages==1.4.1
boto==2.41.0
gunicorn==19.6.0
daphne==1.0.3
asgi_redis==1.0.0
sox==1.2.8

##################################################################
# django modules
##################################################################
django-admin-tools==0.8.0
django-braces==1.10.0
django-celery-beat==1.0.1
django-click==1.2.0
django-compressor==2.0
django-split-settings==0.2.3
django-turbolinks==0.4

##################################################################
# RESTful API
# http://www.django-rest-framework.org/#installation
##################################################################
djangorestframework==3.6.3
markdown==2.6.7
coreapi==2.1.1

##################################################################
# authentication
##################################################################
django-authtools==1.5.0
social-auth-core==1.3.0
social-auth-app-django==1.2.0
django_class_based_auth_views==0.4
django_password_reset==0.9
django-registration==2.2


##################################################################
# task handling
##################################################################
celery[redis]==4.0.2

##################################################################
# other modules
##################################################################
raven==6.0.0


##################################################################
# external modules & patches
##################################################################
-e git+https://github.com/palmbeach-interactive/django-slick-admin@master#egg=django-slick-admin
